LEL-014 wip: adjust database layer

+++ core/database/src/main/java/io/github/faening/lello/core/database/dao/JournalCategoryDao.kt
@@ -52,5 +52,5 @@ interface JournalCategoryDao : JournalCategoryResources<JournalCategoryEntity> {
             LIMIT 1
         """
     )
-    override fun getById(id: Int): Flow<JournalCategoryEntity>?
+    override fun getById(id: Long): Flow<JournalCategoryEntity>?
 }
